get reputation history and load into view

+++ app/src/main/java/com/example/sofuser/entity/response/UserEntity.kt
@@ -20,5 +20,5 @@ data class UserEntity(
     val location: String?,
 
     @SerializedName(ApiKey.LAST_ACCESS_DATE)
-    val lastAccessDate: Int?
+    val lastAccessDate: Long?
 )
